LayoutTests/editing/input/ios/typing-with-inline-predictions-expected.txt

CONSOLE MESSAGE: Typing 't'
CONSOLE MESSAGE: Typing 'o'
CONSOLE MESSAGE: Typing ' '
CONSOLE MESSAGE: Typing 'w'
CONSOLE MESSAGE: Typing 'h'
CONSOLE MESSAGE: Typing 'o'
CONSOLE MESSAGE: Typing 'm'
CONSOLE MESSAGE: Typing ' '
CONSOLE MESSAGE: Typing 'i'
CONSOLE MESSAGE: Typing 't'
CONSOLE MESSAGE: Typing ' '
CONSOLE MESSAGE: Typing 'm'
CONSOLE MESSAGE: Typing 'a'
CONSOLE MESSAGE: Typing 'y'
CONSOLE MESSAGE: Typing ' '
CONSOLE MESSAGE: Typing 'c'
CONSOLE MESSAGE: Typing 'o'
CONSOLE MESSAGE: Typing 'n'
CONSOLE MESSAGE: Typing 'c'
CONSOLE MESSAGE: Typing 'e'
CONSOLE MESSAGE: Typing 'r'
CONSOLE MESSAGE: Typing 'n'
This test verifies that typing with inline predictions does not cause the keyboard to incorrectly autocapitalize words. This test requires WebKitTestRunner.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS isAutoShifted is false
PASS input.value is "To whom it may concern"
PASS successfullyParsed is true

TEST COMPLETE
